rename types->schemas
